Fixed casting error

+++ flint-core/src/util/plugin/mod.rs
@@ -309,7 +309,7 @@ fn toml_value_to_lua_value<'lua>(
 ) -> Result<Value, mlua::Error> {
     match value {
         toml::Value::String(s) => Ok(Value::String(lua.create_string(s)?)),
-        toml::Value::Integer(i) => Ok(Value::Integer(*i)),
+        toml::Value::Integer(i) => Ok(Value::Integer(*i as i32)),
         toml::Value::Float(f) => Ok(Value::Number(*f)),
         toml::Value::Boolean(b) => Ok(Value::Boolean(*b)),
         toml::Value::Datetime(dt) => Ok(Value::String(lua.create_string(&dt.to_string())?)),
